Fixed AO not working when SSR was enabled
Lighting is now composited in two passes:

Lighting pass:
Direct and indirect light is computed.
AO is applyed to the light for the SSR hit texture.
Light without AO is stored in another texture.

In the SSR pass:
AO is applyed back to the stored light.
Then from the previous pass the stored light is substracted.
AO is removed from the stored light.
Now the cube specular is substracted from the stored light.
Either SSR hit or cube specular is applyed back into the stored light.
Finally AO is applyed back into the stored light and the result is added to the previous pass color.

This operation allows AO and SSR to be combined correctly.