Fixed #15
Issue was caused because in the original engine the Y axis was UP and so the
irradiance capture code was designed to keep the lower filter in that axis,
here Z is UP so the lower filter caused the sides to look wrong.

The UP vector has been switched to move back the lower filter to the Z axis instead
and the capture source has been increased to allow for more samples.

Clouds seems to cause the issue to appear in some scenes but this time only the upper
side is affected.